Broke out Avatar component from profile.js to avatar.js in profile folder
